- got rid of the unnessercery intermediate class.
